changes to function, need finishing
